+++ dashboard/.eslintrc.js
@@ -9,7 +9,14 @@ module.exports = {
     },
     rules: {
         //"no-console": process.env.NODE_ENV === "production" ? "error" : "off",
-        "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off"
+        "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off",
+        indent: ["error", 4, { SwitchCase: 1 }],
+        "linebreak-style": [
+            process.env.NODE_ENV === "production" ? "error" : "off",
+            "windows"
+        ],
+        quotes: ["error", "double"],
+        semi: ["error", "always"]
     },
     overrides: [
         {