+++ dashboard/src/error-handler.js
@@ -4,24 +4,34 @@ import Vue from "vue";
 export default function(vueApp) {
     // handle errors
     Vue.config.errorHandler = function(error, vm, info) {
-        if (error.isAxiosError) {
+        if (error.isAxiosError && error.response) {
             let response = error.response;
             switch (response.status) {
                 case 400:
-                    vueApp.$toast.error(response.data);
-                    break;
+                    if (typeof response.data === "string") {
+                        vueApp.$toast.warning(response.data);
+                    } else if (
+                        typeof response.data === "object" &&
+                        response.data.message
+                    ) {
+                        vueApp.$toast.warning(response.data.message);
+                    } else {
+                        vueApp.$toast.error("Unknown error");
+                        console.error(error);
+                    }
+                    return;
 
                 case 401:
                     window.location =
-                        "/signin?return=" + location.pathname + location.search;
+                        process.env.VUE_APP_SIGNIN_URL +
+                        location.pathname +
+                        location.search;
                     return;
 
                 default:
                     vueApp.$toast.error("Request completed with error");
-                    break;
+                    return;
             }
-
-            return;
         }
 
         vueApp.$toast.error("Unknown error occurred");
@@ -44,7 +54,7 @@ export default function(vueApp) {
             console.log(url);
             console.log(line);
             console.log(col);
-            console.log(error);
+            console.error(error);
         }
     };
 
@@ -53,7 +63,7 @@ export default function(vueApp) {
         {
             //if (options.env === "development") {
             console.log("unhandledrejection");
-            console.log(event);
+            console.error(event);
         }
     });
 }