+++ facade/server/lib/api.js
@@ -69,16 +69,18 @@ const call = new Proxy(axios, {
 
 module.exports.users = {
     get: async usernameOrEmail => call.get("user/" + usernameOrEmail),
-    add: async user => call.post("user", user)
+    add: async user => call.post("user", user),
+    canRegister: async () => call.get("users/can-register")
 };
 
 module.exports.photos = {
-    all: () => call.get("/gallery/photos"),
+    all: (sort, skip, limit) => call.get("/gallery/photos", { params: { sort, skip, limit } }),
+    count: () => call.get("/gallery/photos/count"),
     get: id => call.get("/gallery/photo/" + id)
 };
 
 module.exports.photoSets = {
-    notEmpty: () => call.get("/gallery/photosets/not-empty"),
+    all: (sort) => call.get("/gallery/photosets", { params: { sort } }),
     get: code => call.get("/gallery/photoset/" + code),
-    getWithPhotos: code => call.get(`/gallery/photoset/${code}?photos=true`)
+    getWithPhotos: code => call.get(`/gallery/photoset/${code}`, { params: { photos: true } })
 };