+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## **2.1.0** - *2020-04-07*
+* API - `REGISTRATION_MODE` option added (possible values are `open` and `invite`);
+* Facade - show Slick header slideshow only on homepage;
+* Dashboard:
+    * check user authentication before rendering page;
+    * user authentication & registration implemented inside Dashboard;
+* Gallery:
+    * `GALLERY_NEW_PHOTOS_FIRST` option added;
+    * resize photoset cover;
+    * `jimp` support dropped;
+    * refresh photoset's cover after changing it;
+    * view photo set - load the smallest thumbnails by default;
+    * Facade - consider `GALLERY_IMAGE_SIZES` and `GALLERY_PHOTOSET_COVER_SIZES` options in `justifiedGallery`;
+    * image loading adjusted (show blurred thumbnail while loading larger image);
+    * show last `GALLERY_LAST_PHOTOS_COUNT` photos below photo sets;
+    * navigation in `All Photos` photo set;
+    * photo sets order - init;
+* dev proxy - copy dev URL to clipboard;
+* minor changes;
+
 ## **2.0.1** - *2020-04-01*
 * ability to use MongoDB as session storage added;
 * consider `FACADE_PORT`, `DASHBOARD_PORT`, `API_PORT` env vars;